+++ tests/test_04_dxf_high_level_structs/test_413_dxfgroups.py
@@ -1,9 +1,9 @@
-# Copyright (c) 2020, Manfred Moitzi
+# Copyright (c) 2020-2021, Manfred Moitzi
 # License: MIT License
 
 import pytest
 import ezdxf
-
+from ezdxf.audit import Auditor
 
 @pytest.fixture(scope='module')
 def doc():
@@ -22,7 +22,6 @@ def test_new_group(doc):
     assert len(list(group.handles())) == 2
     handle = group[0].dxf.handle
     assert handle in group
-
 
 def test_unique_groups(doc):
     doc.groups.new('test2')
@@ -43,6 +42,38 @@ def test_modify_group(doc):
     ]
     group.extend(e)
     assert len(group) == 4
+
+
+def test_can_not_add_invalid_block_entities(doc):
+    group = doc.groups.new('test4')
+    block = doc.blocks.new("Block4")
+    point = block.add_point((0, 0))
+    with pytest.raises(ezdxf.DXFStructureError):
+        with group.edit_data() as g:
+            g.append(point)
+
+
+def test_can_not_add_invalid_table_entry(doc):
+    group = doc.groups.new('test5')
+    layer = doc.layers.get("0")
+    with pytest.raises(ezdxf.DXFStructureError):
+        with group.edit_data() as g:
+            g.append(layer)
+
+
+def test_audit_filters_invalid_entities(doc):
+    group = doc.groups.new('test6')
+    msp = doc.modelspace()
+    block = doc.blocks.new("Block6")
+    point1 = block.add_point((0, 0))  # invalid BLOCK entity
+    point2 = msp.add_point((0, 0))  # valid model space entity ...
+    point2.destroy()  # ... but destroyed
+    layer = doc.layers.get("0")  # invalid table entry
+
+    group.extend([point1, point2, layer])
+    auditor = Auditor(doc)
+    group.audit(auditor)
+    assert len(group) == 0
 
 
 if __name__ == '__main__':